feat(nav): raise the estimator to the IMU's own 1014.2 Hz rate (T15)
The rate change and only the rate change (docs/REFACTORING_PLAN.md
§3.4/§3.6), gated on T14's prerequisites and on §3.5's measured
execMaxUs = 170 us / 300 us budget:

- ImuInt.c: retarget SRC_SCUERU0 to CPU1 and IFX_INTERRUPT's vectabNum
  to 1, matching -- the T12 trap (docs/ILLD_NOTES.md §3 T15), verified
  in the .map (imuDrdyIsr lands in int_tab_tc1 @ 0x805fed40) and in the
  generated .src (a real DSYNC in both the ISR and ImuInt_init()).
- ImuEdge.h/ImuEdgePlace.c (new): {ticks, seq} producer/consumer object
  in the LMU shared block (0xB00F0500, the fourth crossing), single
  writer imuDrdyIsr. Kept separate from the existing g_imuDrdyLastTicks/
  g_imuDrdyCount bring-up globals, which stay plain (tools/xcp_read.py
  and tools/imu_int_stats.py read them by name; folding them into a
  struct would break that).
- NavTask.c: NavTask_step takes dt from the edge timestamps instead of
  SysTime_getTimeElapsedS(), consumes at most one pending edge per
  dispatch, and returns almost immediately otherwise. Missed edges
  (seq delta > 1) are counted into g_imuDrdyMissedEdges. A bounded
  fallback (NAVTASK_NO_EDGE_TIMEOUT_S = 20 ms, matching the old 50 Hz
  task's own worst-case freshness) still probes Icm42688_read() when no
  edge has arrived at all, so a fully disconnected sensor is still
  noticed and its hot-plug recovery still runs -- a pure newSample gate
  would otherwise never call it again once its own edges stopped.
- Cpu1_Main.c: NavTask_step registered first, at SCHED_US(500) (2 kHz
  poll against the ~1014 Hz sensor), Task_LedToggle after it.
- Icm42688.c: ICM42688_RECOVERY_PERIOD's comment corrected for the new
  call cadence (no constant change).
- GnssM9N.c: one incidental fix -- an unrelated header touch forced this
  file's first rebuild in a while and surfaced a pre-existing implicit-
  truncation warning (g_ubx_payload_len uint16 -> uint8 param); added an
  explicit, already-bounds-checked cast to keep the build at 0 warnings.

test/test_imuedge.c (new): the snapshot/torn-read protocol, same
limitation test_navstate.c already documents (retry needs a real
concurrent writer, untestable single-threaded). test_navtask.c: two new
stubs for the additional linked symbols.

+++ src/bsw/GnssM9N.c
@@ -679,7 +679,17 @@ boolean GnssM9N_read(GnssM9N_Sample *sample){
 
                     if (g_len >= (g_ubx_payload_len + 6u))
                     {
-                        if (gnss_ubx_decode(g_buffer, g_len, g_ubx_payload_len, &g_nav) != TRUE)
+                        /* g_ubx_payload_len is uint16 (ubx_rd_u16, above) but
+                         * bounded to <= GNSS_UBX_MAX_RX_PAYLOAD (94) on every
+                         * path that reaches this call: g_len is uint8 and the
+                         * guard above requires g_len >= payload_len + 6, so a
+                         * payload_len this call ever sees cannot exceed
+                         * g_len's own max (GNSSM9N_BUFFER_SIZE). Explicit cast
+                         * -- found while an unrelated header touch forced
+                         * this file's first rebuild in a while and surfaced
+                         * the implicit-truncation warning; not part of any
+                         * planned change here. */
+                        if (gnss_ubx_decode(g_buffer, g_len, (uint8)g_ubx_payload_len, &g_nav) != TRUE)
                         {
                             g_errors++;
                         }
